Fix remapping endpoint and move/fix integration tests around
